// frontend/src/core/api/generated/model/publicPlanView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */
import type { PlanFeatureView } from './planFeatureView'

export interface PublicPlanView {
  id: string
  code: string
  name: string
  /** @nullable */
  tagline?: string | null
  /** @nullable */
  monthlyPrice?: number | null
  currency: string
  /** @nullable */
  maxUsers?: number | null
  /** @nullable */
  maxSites?: number | null
  highlighted: boolean
  features: PlanFeatureView[]
}